+++ orm/src/main/java/org/hibernate/tool/ant/Hbm2DocExporterTask.java
@@ -4,8 +4,8 @@
  */
 package org.hibernate.tool.ant;
 
+import org.hibernate.tool.api.export.Exporter;
 import org.hibernate.tool.hbm2x.DocExporter;
-import org.hibernate.tool.hbm2x.Exporter;
 
 public class Hbm2DocExporterTask extends ExporterTask {
