+++ orm/src/main/java/org/hibernate/tool/ant/HbmLintExporterTask.java
@@ -1,6 +1,6 @@
 package org.hibernate.tool.ant;
 
-import org.hibernate.tool.hbm2x.Exporter;
+import org.hibernate.tool.api.export.Exporter;
 import org.hibernate.tool.hbm2x.HbmLintExporter;
 
 public class HbmLintExporterTask extends ExporterTask {
