+++ orm/src/main/java/org/hibernate/tool/hbm2x/AbstractExporter.java
@@ -12,6 +12,7 @@ import org.hibernate.boot.Metadata;
 import org.hibernate.internal.util.ReflectHelper;
 import org.hibernate.internal.util.StringHelper;
 import org.hibernate.tool.api.export.ArtifactCollector;
+import org.hibernate.tool.api.export.Exporter;
 import org.hibernate.tool.api.metadata.MetadataDescriptor;
 import org.hibernate.tool.internal.export.DefaultArtifactCollector;
 import org.slf4j.Logger;
